refactoring job master classes
